backend/requirements.txt
psycopg2==2.9.5
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Security==3.0.0
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
Flask-Script==2.0.5
Flask-Limiter==2.8.1
flask-apispec==0.11.4
marshmallow==3.19.0
email-validator==1.3.0
redis==4.4.0
gunicorn==20.1.0
gevent==22.10.2
python-dotenv==0.21.0
pydantic==1.10.2
bcrypt==4.0.1
rauth==0.7.3
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
user-agents==2.2.0
python-logstash==0.4.8
pytz==2023.3